Refactoring and general improvements to TFL example.
Now able to run in parallel without multiple runners (runners are generated on the fly) and allure reporting will now shows each scenario as a step with substeps for their operations.

Still TODO: Screenshotting on failure won't work, as the default screenshot on fail fires at the end of a test (which in this case means the end of a whole feature) - because we close the browser at the end of the scenario, by the time the screenshotter gets round to it we've killed the driver already.

Look to make tags specification better - it should be possible to run with -Dthreads=2 -Dtags=something, but this isn't going to work just yet